requirements.txt
fastapi
uvicorn
watchdog
chromadb
sentence-transformers
python-dotenv
google-generativeai